Make function available outside source file
